requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
ansible==8.7.0
ansible-core==2.15.8
cffi==1.16.0
contourpy==1.0.7
cryptography==41.0.7
cycler==0.11.0
fonttools==4.39.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
importlib-resources==5.0.7
Jinja2==3.1.2
kiwisolver==1.4.4
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
resolvelib==1.0.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
zipp==3.15.0
